Cricbuzz_Live_Stats/requirements.txt
# requirements.txt
streamlit==1.28.1
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pandas==2.0.3